#181: renamed VT writer/reader for style consistency
